refactor: remove mutation resolver
